backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2

# Database
pymongo==4.6.0
motor==3.3.2

# Data Processing
numpy==1.24.3
pandas==2.1.3

# External APIs
requests==2.31.0
feedparser==6.0.10
python-dotenv==1.0.0

# Scheduling
schedule==1.2.0

# ML/Data Science (future use)
prophet==1.1.5
scikit-learn==1.3.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

# Twilio for SMS support and i18n for multi-language
twilio==8.10.0
python-i18n==0.3.9
